pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Leonard Schuler", email = "31131665+LeonardSchuler@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.62",
    "fastapi[standard]>=0.120.2",
    "python-dotenv>=1.2.1",
]

[project.scripts]
rag = "rag.app.main:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]